pyproject.toml
[project]
name = "ta-flnks-pe"
version = "0.1.0"
description = ""
authors = [
    {name = "Renato Ramos Silva",email = "renatoramossilva@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
uvicorn = "^0.34.2"
fastapi = "^0.115.12"
playwright = "^1.52.0"
bindl-lib = {git = "https://github.com/renatoramossilva/bindl-lib.git", branch = "add-aws-wrapper"}
black = "^25.1.0"
ruff = "^0.12.0"
mypy = "^1.15.0"
pre-commit = "^4.2.0"
pytest = "^8.3.5"
boto3 = "^1.38.36"
docker = "^7.1.0"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.26.0"
httpx = "^0.28.1"

[tool.ruff]
# Defaults: https://beta.ruff.rs/docs/configuration/#using-pyprojecttoml
target-version = "py310"
fix = true
show-fixes = true
respect-gitignore = true

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "ANN204",  # Missing return type annotation for special method `__init__`
    "E501",    # Line too long
    "D107",    # Missing docstring in `__init__`
    "TRY400",  # Use `logging.exception` instead of `logging.error`
    "D104"     # Missing docstring in public package
]

# Config to avoid pytest warning
# eprecationWarning: The configuration option
# "asyncio_default_fixture_loop_scope" is unset.
[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
